feat: add advanced screen
